grafana-ui: Adopt lazy DnD in smaller consumers and defer value mappings and job status
